refactored database, added players screens on frontend

+++ fbcoach-backend/fbcoach/src/main/java/hr/fer/fbcoach/service/impl/PlayerServiceJpa.java
@@ -1,8 +1,10 @@
 package hr.fer.fbcoach.service.impl;
 
 import hr.fer.fbcoach.model.Player;
+import hr.fer.fbcoach.model.Team;
 import hr.fer.fbcoach.model.exception.ResourceNotFoundException;
 import hr.fer.fbcoach.repository.PlayerRepository;
+import hr.fer.fbcoach.repository.TeamRepository;
 import hr.fer.fbcoach.service.PlayerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,9 +17,14 @@ import java.util.Optional;
 public class PlayerServiceJpa implements PlayerService {
 
     private final PlayerRepository playerRepository;
+    private final TeamRepository teamRepository;
 
     public List<Player> getAllPlayers() {
         return playerRepository.findAll();
+    }
+
+    public List<Player> getAllPlayersByTeamId(Long teamId) {
+        return playerRepository.findByTeams_IdTeam(teamId);
     }
 
     public Optional<Player> getPlayerById(Long id) {
@@ -41,6 +48,54 @@ public class PlayerServiceJpa implements PlayerService {
     }
 
     public void deletePlayer(Long id) {
-        playerRepository.deleteById(id);
+        Optional<Player> playerOptional = playerRepository.findById(id);
+        if (playerOptional.isPresent()) {
+            Player player = playerOptional.get();
+            List<Team> teams = player.getTeams();
+
+            // Remove player from all teams
+            for (Team team : teams) {
+                team.getPlayers().remove(player);
+            }
+
+            // Save all teams to update the changes in the database
+            teamRepository.saveAll(teams);
+
+            // Delete the player from the repository
+            playerRepository.deleteById(id);
+        }
     }
+
+    public void removePlayerFromTeam(Long playerId, Long teamId) {
+        Optional<Player> playerOptional = playerRepository.findById(playerId);
+        Optional<Team> teamOptional = teamRepository.findById(teamId);
+
+        if (playerOptional.isPresent() && teamOptional.isPresent()) {
+            Player player = playerOptional.get();
+            Team team = teamOptional.get();
+
+            player.getTeams().remove(team);
+            team.getPlayers().remove(player);
+
+            playerRepository.save(player);
+        }
+    }
+
+    public void addPlayerToTeam(Long playerId, Long teamId) {
+        Optional<Player> playerOptional = playerRepository.findById(playerId);
+        Optional<Team> teamOptional = teamRepository.findById(teamId);
+
+        if (playerOptional.isPresent() && teamOptional.isPresent()) {
+            Player player = playerOptional.get();
+            Team team = teamOptional.get();
+
+            player.getTeams().add(team);
+            team.getPlayers().add(player);
+
+            playerRepository.save(player);
+            teamRepository.save(team);
+        }
+    }
+
+
 }
